+++ src/storage/slots.rs
@@ -1,12 +1,12 @@
 use {
-    crate::metrics,
+    crate::{
+        metrics,
+        util::{HashMap, HashSet},
+    },
     solana_sdk::{clock::Slot, commitment_config::CommitmentLevel},
-    std::{
-        collections::{HashMap, HashSet},
-        sync::{
-            Arc, Mutex,
-            atomic::{AtomicU64, Ordering},
-        },
+    std::sync::{
+        Arc, Mutex,
+        atomic::{AtomicBool, AtomicU64, Ordering},
     },
 };
 
@@ -16,6 +16,7 @@ pub struct StoredSlots {
     confirmed: Arc<AtomicU64>,
     finalized: Arc<AtomicU64>,
     first_available: Arc<AtomicU64>,
+    max_recent_blockhashes: Arc<AtomicBool>,
 }
 
 impl Default for StoredSlots {
@@ -25,6 +26,7 @@ impl Default for StoredSlots {
             confirmed: Arc::new(AtomicU64::new(u64::MIN)),
             finalized: Arc::new(AtomicU64::new(u64::MIN)),
             first_available: Arc::new(AtomicU64::new(u64::MAX)),
+            max_recent_blockhashes: Arc::new(AtomicBool::new(false)),
         }
     }
 }
@@ -35,6 +37,7 @@ impl StoredSlots {
             && self.confirmed_load() != u64::MIN
             && self.finalized_load() != u64::MIN
             && self.first_available_load() != u64::MAX
+            && self.max_recent_blockhashes.load(Ordering::Relaxed)
     }
 
     pub fn processed_load(&self) -> Slot {
@@ -76,30 +79,74 @@ impl StoredSlots {
 }
 
 #[derive(Debug, Clone)]
-pub struct StoredConfirmedSlot {
+pub struct StoredSlotsRead {
     stored_slots: StoredSlots,
-    slots: Arc<Mutex<HashMap<Slot, HashSet<usize>>>>,
+    slots_confirmed: Arc<Mutex<HashMap<Slot, HashSet<usize>>>>,
+    slots_finalized: Arc<Mutex<HashMap<Slot, HashSet<usize>>>>,
+    max_recent_blockhashes: Arc<Mutex<usize>>,
+    max_recent_blockhashes_ready: bool,
     total_readers: usize,
 }
 
-impl StoredConfirmedSlot {
+impl StoredSlotsRead {
     pub fn new(stored_slots: StoredSlots, total_readers: usize) -> Self {
         Self {
             stored_slots,
-            slots: Arc::default(),
+            slots_confirmed: Arc::default(),
+            slots_finalized: Arc::default(),
+            max_recent_blockhashes: Arc::default(),
+            max_recent_blockhashes_ready: false,
             total_readers,
         }
     }
 
-    pub fn set_confirmed(&self, index: usize, slot: Slot) {
-        let mut lock = self.slots.lock().expect("unpanicked mutex");
+    fn set(
+        &self,
+        map: &Arc<Mutex<HashMap<Slot, HashSet<usize>>>>,
+        index: usize,
+        slot: Slot,
+    ) -> bool {
+        let mut lock = map.lock().expect("unpanicked mutex");
 
         let entry = lock.entry(slot).or_default();
         entry.insert(index);
 
         if entry.len() == self.total_readers {
-            self.stored_slots.confirmed_store(slot);
             lock.remove(&slot);
+            true
+        } else {
+            false
+        }
+    }
+
+    pub fn set_confirmed(&self, index: usize, slot: Slot) {
+        if self.set(&self.slots_confirmed, index, slot) {
+            self.stored_slots.confirmed_store(slot);
+        }
+    }
+
+    pub fn set_finalized(&self, index: usize, slot: Slot) {
+        if self.set(&self.slots_finalized, index, slot) {
+            self.stored_slots.finalized_store(slot);
+        }
+    }
+
+    // max recent blockhashes
+    pub fn set_ready(&mut self, ready: bool) {
+        if ready && !self.max_recent_blockhashes_ready {
+            self.max_recent_blockhashes_ready = true;
+
+            let mut lock = self
+                .max_recent_blockhashes
+                .lock()
+                .expect("unpanicked mutex");
+            *lock += 1;
+
+            if *lock == self.total_readers {
+                self.stored_slots
+                    .max_recent_blockhashes
+                    .store(true, Ordering::Relaxed);
+            }
         }
     }
 }