+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 - rpc: use confirmed during sync ([#6](https://github.com/lamports-dev/alpamayo/pull/6))
 - storage: remove transaction index on slot remove ([#8](https://github.com/lamports-dev/alpamayo/pull/8))
 - storage: set confirmed/finalized on first stream messages ([#17](https://github.com/lamports-dev/alpamayo/pull/17))
+- rpc: load recent blocks on startup ([#23](https://github.com/lamports-dev/alpamayo/pull/23))
 
 ### Features